backend_python/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
